Tweaks to get AIX to build Doxygen
CMake checks for libiconv, but the check in Cmake was failing becuse it was
using the wrong library, to fix that a profile for building on AIX was added,
it sets the environment variable LDFLAGS so that it is added when linking when
used.
The Doxygen project's CMakeLists.txt needed to be patched so  it would not
fail when it couldn't find libiconv.
Doxygen code had many issues building on AIX with gcc, needed to add a header
file, undefine TRUE and FALSE, and also to set the flag -mcmodel=large or the
assembler throws errors.
Remove the build from archive option, and always build from source.
Remove the cmake dependency from conan, Linux packages don't run when this
is used because outside DL, binaries are build on machine with later glibc
versions.  Getting Cmake through conan requires Cmake to build, and in this
case was using a newer cmake to build the conan package in most cases.
The one case where this is nice is when a package is  already built and
sitting on our artifactory, then it doesn't require cmake to build cmake.